Make Separate class for PopupView to reuse
